+++ CHANGELOG.md
@@ -36,11 +36,12 @@
     - Relies on N-1 system for when matching KDK is not present
   - Delete unused KDKs in `/Library/Developer/KDKs` during root patching
   - Resolve Power Management support for Ivy Bridge and older
+- Add work-around to Catalyst Buttons not responding on non-Metal in macOS Monterey
 - Increment Binaries:
   - OpenCorePkg 0.8.5 release
   - Lilu 1.6.2 - release
   - FeatureUnlock 1.0.9 release
-  - PatcherSupportPkg 0.6.8 - release
+  - PatcherSupportPkg 0.6.9 - release
   - BrcmPatchRAM 2.6.4 - release
   - AutoPkgInstaller 1.0.1 - release
   - CryptexFixup 1.0.0 - release